+++ src/main/java/com/yuuko/core/Configuration.java
@@ -39,7 +39,7 @@ import java.util.*;
 public class Configuration {
     private static final Logger log = LoggerFactory.getLogger(Configuration.class);
 
-    public static final String VERSION = "2019-07-14";
+    public static final String VERSION = "2019-07-18";
     public static String AUTHOR;
     public static String AUTHOR_WEBSITE;
     public static String SUPPORT_GUILD;